+++ app/src/main/java/com/oracle/visualize/data/mapper/AuthMappers.kt
@@ -17,49 +17,3 @@ fun FirebaseUser.toDomain(): AuthUser = AuthUser(
     uid = uid,
     email = email ?: ""
 )
-
-fun VisualizationDto.toDomain(): Visualization = Visualization(
-    id = id,
-    authorID = authorID,
-    title = title,
-    configJSON = configJSON,
-    sharedWithUsers = sharedWithUsers,
-    sharedWithTeams = sharedWithTeams,
-    createdAt = createdAt,
-    comments = comments.map { it.toDomain() }
-)
-
-fun TeamDto.toDomain(): Team = Team(
-    id = id,
-    memberIDs = memberIDs,
-    name = name,
-    ownerID = ownerID
-)
-
-fun UserDto.toDomain(): User = User(
-    id = id,
-    userType = userType,
-    email = email,
-    userName = userName,
-    profilePictureUrl = profilePictureUrl,
-    themePreference = themePreference,
-    chartTheme = chartTheme,
-    notificationsEnabled = notificationsEnabled,
-    hiddenVisualizations = hiddenVisualizations
-)
-
-fun CommentDto.toDomain(): Comment = Comment(
-    id = id,
-    authorID = authorID,
-    content = content,
-    createdAt = createdAt,
-    imageUrl = imageUrl,
-    threads = threads.map { it.toDomain() }
-)
-
-fun ThreadDto.toDomain(): Thread = Thread(
-    id = id,
-    authorID = authorID,
-    content = content,
-    timestamp = timestamp
-)
